[biosilico-creatures]Parsing Creatiures 1 Genetics Upgrades (type/subtype of genes)
